port80: remove test
KVM has not passed port 0x80 directly to the hardware for three years
(commit d59d51f08801, "KVM: VMX: remove I/O port 0x80 bypass on Intel
hosts", 2017-12-05) so the port80 test is a useless duplicate of the
outl_to_pmtimer vmexit test, without the reporting of how long the
access takes and without adaptive choice of the number of iterations.
Remove it.

Reviewed-by: Jim Mattson <jmattson@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>