arm/arm64: gic: Remove unnecessary synchronization with stats_reset()
The GICv3 driver executes a DSB barrier before sending an IPI, which
ensures that memory accesses have completed. This removes the need to
enforce ordering with respect to stats_reset() in the IPI handler.

For GICv2, the wmb() from writel(), executed when triggering an IPI,
coupled with the rmb() from readl() when handling the IPI will enforce the
desired memory ordering.

Reviewed-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Andrew Jones <drjones@redhat.com>