arm/arm64: Avoid calling cpumask_test_cpu for CPUs above nr_cpus
Prior to this change, a call of cpumask_next(cpu, mask) where
cpu=nr_cpus - 1 (assuming all cpus are enumerated in the range
0..nr_cpus - 1) would make an out-of-bounds access to the mask.
In many cases, this is still a valid memory location due the
implementation of cpumask_t, however, in certain configurations
(for example, nr_cpus == sizeof(long)) this would cause an access
outside the bounds of the mask too.

This patch changes the way we guard calls to cpumask_test_cpu() in
cpumask_next() to avoid the above condition. A following change adds
assertions to catch out-of-bounds accesses to cpumask_t.

Reviewed-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>
Signed-off-by: Andrew Jones <drjones@redhat.com>