arm/arm64: Read system registers to get the state of the MMU
When we are in EL1 we can directly tell if the local cpu's MMU is on
by reading a system register (SCTRL/SCTRL_EL1). In EL0, we use the
relevant cpumask. This way we don't have to rely on the cpu id in
thread_info when we are in setup executing in EL1.

In addition, this change:
 * Removes mmu_disabled_cpu_count as it is no longer necessary and
   assumed that calls to mmu_mark_enabled()/mmu_mark_disabled() were
   serialized. This is currently true but a future change could
   easily break that assumption.
 * Changes mmu_disabled_mask to mmu_enabled_mask and inverts the
   logic to track in a more intuitive way that all CPUs start with
   the MMU OFF and at some point, we turn them ON.

Reviewed-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>
[ Applied some code cleanups suggested by Alexandru Elisei ]
Signed-off-by: Andrew Jones <drjones@redhat.com>